[ADD] 추가적인 API model 추가 (#100)
