+++ Maddori.Apple/Maddori.Apple/Network/Response/AllCertainTypeFeedBackResponse.swift
@@ -1,0 +1,12 @@
+//
+//  AllCertainTypeFeedBackResponse.swift
+//  Maddori.Apple
+//
+//  Created by Mingwan Choi on 2022/11/16.
+//
+
+import Foundation
+
+struct AllCertainTypeFeedBackResponse: Decodable {
+    let feedback: [FeedBackResponse]
+}
