+++ Maddori.Apple/Maddori.Apple/Network/Response/AllReflectionResponse.swift
@@ -1,0 +1,12 @@
+//
+//  AllReflectionResponse.swift
+//  Maddori.Apple
+//
+//  Created by Mingwan Choi on 2022/11/16.
+//
+
+import Foundation
+
+struct AllReflectionResponse: Decodable {
+    let reflection: [ReflectionResponse]?
+}
