+++ Maddori.Apple/Maddori.Apple/Network/Response/ReflectionInfoResponse.swift
@@ -1,0 +1,12 @@
+//
+//  ReflectionInfoResponse.swift
+//  Maddori.Apple
+//
+//  Created by Mingwan Choi on 2022/11/16.
+//
+
+import Foundation
+
+struct ReflectionInfoResponse: Decodable {
+    let reflection: ReflectionResponse?
+}
